move sender logic into RTCEngine
