+++ src/room/participant/LocalParticipant.ts
@@ -20,10 +20,7 @@ import RTCEngine from '../RTCEngine';
 import LocalAudioTrack from '../track/LocalAudioTrack';
 import LocalTrack from '../track/LocalTrack';
 import LocalTrackPublication from '../track/LocalTrackPublication';
-import LocalVideoTrack, {
-  SimulcastTrackInfo,
-  videoLayersFromEncodings,
-} from '../track/LocalVideoTrack';
+import LocalVideoTrack, { videoLayersFromEncodings } from '../track/LocalVideoTrack';
 import {
   AudioCaptureOptions,
   BackupVideoCodec,
@@ -33,11 +30,10 @@ import {
   ScreenSharePresets,
   TrackPublishOptions,
   VideoCaptureOptions,
-  VideoCodec,
 } from '../track/options';
 import { Track } from '../track/Track';
 import { constraintsForOptions, mergeDefaultOptions } from '../track/utils';
-import { isFireFox, isWeb, supportsAddTrack, supportsTransceiver } from '../utils';
+import { isFireFox, isWeb } from '../utils';
 import Participant from './Participant';
 import { ParticipantTrackPermission, trackPermissionToProto } from './ParticipantTrackPermission';
 import {
@@ -560,15 +556,6 @@ export default class LocalParticipant extends Participant {
     if (encodings) {
       transceiverInit.sendEncodings = encodings;
     }
-    // addTransceiver for react-native is async. web is synchronous, but await won't effect it.
-    const transceiver = await this.engine.publisher.pc.addTransceiver(
-      track.mediaStreamTrack,
-      transceiverInit,
-    );
-    if (track.kind === Track.Kind.Video && opts.videoCodec) {
-      this.setPreferredCodec(transceiver, track.kind, opts.videoCodec);
-      track.codec = opts.videoCodec;
-    }
 
     if (track.codec === 'av1' && encodings && encodings[0]?.maxBitrate) {
       this.engine.publisher.setTrackCodecBitrate(
@@ -581,7 +568,7 @@ export default class LocalParticipant extends Participant {
     this.engine.negotiate();
 
     // store RTPSender
-    track.sender = await this.createSender(track, opts, encodings);
+    track.sender = await this.engine.createSender(track, opts, encodings);
     if (track instanceof LocalVideoTrack) {
       track.startMonitor(this.engine.client);
     } else if (track instanceof LocalAudioTrack) {
@@ -593,97 +580,6 @@ export default class LocalParticipant extends Participant {
     // send event for publication
     this.emit(ParticipantEvent.LocalTrackPublished, publication);
     return publication;
-  }
-
-  async createSender(
-    track: LocalTrack,
-    opts: TrackPublishOptions,
-    encodings?: RTCRtpEncodingParameters[],
-  ) {
-    if (supportsTransceiver()) {
-      return this.createSenderTransceiver(track, opts, encodings);
-    }
-    if (supportsAddTrack()) {
-      console.log('using add-track fallback');
-      return this.getTrackSender(track.mediaStreamTrack);
-    }
-    throw new UnexpectedConnectionState('Required webRTC APIs not supported on this device');
-  }
-
-  async createSenderTransceiver(
-    track: LocalTrack,
-    opts: TrackPublishOptions,
-    encodings?: RTCRtpEncodingParameters[],
-  ) {
-    if (!this.engine.publisher) {
-      throw new UnexpectedConnectionState('publisher is closed');
-    }
-
-    const transceiverInit: RTCRtpTransceiverInit = { direction: 'sendonly' };
-    if (encodings) {
-      transceiverInit.sendEncodings = encodings;
-    }
-    // addTransceiver for react-native is async. web is synchronous, but await won't effect it.
-    const transceiver = await this.engine.publisher.pc.addTransceiver(
-      track.mediaStreamTrack,
-      transceiverInit,
-    );
-    if (track.kind === Track.Kind.Video && opts.videoCodec) {
-      this.setPreferredCodec(transceiver, track.kind, opts.videoCodec);
-      track.codec = opts.videoCodec;
-    }
-    return transceiver.sender;
-  }
-
-  async getSimulcastTransceiverSender(
-    track: LocalVideoTrack,
-    simulcastTrack: SimulcastTrackInfo,
-    opts: TrackPublishOptions,
-    encodings?: RTCRtpEncodingParameters[],
-  ) {
-    if (!this.engine.publisher) {
-      throw new UnexpectedConnectionState('publisher is closed');
-    }
-    const transceiverInit: RTCRtpTransceiverInit = { direction: 'sendonly' };
-    if (encodings) {
-      transceiverInit.sendEncodings = encodings;
-    }
-    // addTransceiver for react-native is async. web is synchronous, but await won't effect it.
-    const transceiver = await this.engine.publisher.pc.addTransceiver(
-      simulcastTrack.mediaStreamTrack,
-      transceiverInit,
-    );
-    if (!opts.videoCodec) {
-      return;
-    }
-    this.setPreferredCodec(transceiver, track.kind, opts.videoCodec);
-    track.setSimulcastTrackSender(opts.videoCodec, transceiver.sender);
-  }
-
-  async getTrackSender(track: MediaStreamTrack) {
-    if (!this.engine.publisher) {
-      throw new UnexpectedConnectionState('publisher is closed');
-    }
-    return this.engine.publisher.pc.addTrack(track);
-  }
-
-  async setupSimulcastSender(
-    track: LocalVideoTrack,
-    simulcastTrack: SimulcastTrackInfo,
-    opts: TrackPublishOptions,
-    encodings?: RTCRtpEncodingParameters[],
-  ) {
-    // store RTCRtpSender
-    // @ts-ignore
-    if (supportsTransceiver()) {
-      return this.getSimulcastTransceiverSender(track, simulcastTrack, opts, encodings);
-    }
-    if (supportsAddTrack()) {
-      console.log('using add-track');
-      return this.getTrackSender(track.mediaStreamTrack);
-    }
-
-    throw new UnexpectedConnectionState('Cannot stream on this device');
   }
 
   /** @internal
@@ -747,15 +643,11 @@ export default class LocalParticipant extends Participant {
 
     const ti = await this.engine.addTrack(req);
 
-    if (!this.engine.publisher) {
-      throw new UnexpectedConnectionState('publisher is closed');
-    }
     const transceiverInit: RTCRtpTransceiverInit = { direction: 'sendonly' };
     if (encodings) {
       transceiverInit.sendEncodings = encodings;
     }
-    // addTransceiver for react-native is async. web is synchronous, but await won't effect it.
-    await this.setupSimulcastSender(track, simulcastTrack, opts, encodings);
+    await this.engine.setupSimulcastSender(track, simulcastTrack, opts, encodings);
 
     this.engine.negotiate();
     log.debug(`published ${videoCodec} for track ${track.sid}`, { encodings, trackInfo: ti });
@@ -1088,50 +980,6 @@ export default class LocalParticipant extends Participant {
       }
     });
     return publication;
-  }
-
-  private setPreferredCodec(
-    transceiver: RTCRtpTransceiver,
-    kind: Track.Kind,
-    videoCodec: VideoCodec,
-  ) {
-    if (!('getCapabilities' in RTCRtpSender)) {
-      return;
-    }
-    const cap = RTCRtpSender.getCapabilities(kind);
-    if (!cap) return;
-    log.debug('get capabilities', cap);
-    const matched: RTCRtpCodecCapability[] = [];
-    const partialMatched: RTCRtpCodecCapability[] = [];
-    const unmatched: RTCRtpCodecCapability[] = [];
-    cap.codecs.forEach((c) => {
-      const codec = c.mimeType.toLowerCase();
-      if (codec === 'audio/opus') {
-        matched.push(c);
-        return;
-      }
-      const matchesVideoCodec = codec === `video/${videoCodec}`;
-      if (!matchesVideoCodec) {
-        unmatched.push(c);
-        return;
-      }
-      // for h264 codecs that have sdpFmtpLine available, use only if the
-      // profile-level-id is 42e01f for cross-browser compatibility
-      if (videoCodec === 'h264') {
-        if (c.sdpFmtpLine && c.sdpFmtpLine.includes('profile-level-id=42e01f')) {
-          matched.push(c);
-        } else {
-          partialMatched.push(c);
-        }
-        return;
-      }
-
-      matched.push(c);
-    });
-
-    if ('setCodecPreferences' in transceiver) {
-      transceiver.setCodecPreferences(matched.concat(partialMatched, unmatched));
-    }
   }
 
   /** @internal */
